+++ app/src/at-quynhho/java/com/asiantech/CheckEmailActivity.kt
@@ -1,12 +1,13 @@
-package com.asiantech.summer
+package com.asiantech
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.asiantech.summer.R
 
-class SignUpActivity : AppCompatActivity() {
+class CheckEmailActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sign_up)
+        setContentView(R.layout.activity_check_email)
     }
 }
