+++ app/src/at-quynhho/res/drawable/ic_background_edittop.xml
