+++ app/src/at-quynhho/res/layout/activity_check_email.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.asiantech.summer.com.asiantech.CheckEmailActivity">
+    tools:context="com.asiantech.CheckEmailActivity">
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
