+++ app/src/at-quynhho/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.asiantech.summer.com.asiantech.SignUpActivity">
+    tools:context="com.asiantech.SignUpActivity">
 <androidx.constraintlayout.widget.ConstraintLayout
 
     android:layout_width="match_parent"
