+++ app/src/at-quynhho/res/values/colors.xml
@@ -3,4 +3,12 @@
     <color name="colorPrimary">#6200EE</color>
     <color name="colorPrimaryDark">#3700B3</color>
     <color name="colorAccent">#03DAC5</color>
+    <color name="colorBlue">#1a237e</color>
+    <color name="colorBlack">#212121</color>
+    <color name="colorFb">#8c9eff</color>
+    <color name="colorTw">#90caf9</color>
+    <color name="colorYt">#ffccbc</color>
+    <color name="colorGreyLight">#eeeeee</color>
+    <color name="colorGrey">#bdbdbd</color>
+    <color name="colorWhite">#ffffff</color>
 </resources>
