+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SignUpActivity"/>
+        <activity android:name="com.asiantech.SignUpActivity"/>
         <activity android:name=".Activity" />
-        <activity android:name=".CheckEmailActivity" />
-        <activity android:name=".LoginActivity">
+        <activity android:name="com.asiantech.CheckEmailActivity" />
+        <activity android:name="com.asiantech.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
