+++ app/src/main/res/values/strings.xml
@@ -1,22 +1,4 @@
 <resources>
     <string name="app_name">Intern20Summer</string>
-    <string name="title_activity_login">Sign in</string>
-    <string name="prompt_email">Email</string>
-    <string name="prompt_password">Password</string>
-    <string name="action_sign_in">Sign in or register</string>
-    <string name="action_sign_in_short">Sign in</string>
-    <string name="welcome">"Welcome !"</string>
-    <string name="invalid_username">Not a valid username</string>
-    <string name="invalid_password">Password must be >5 characters</string>
-    <string name="login_failed">"Login failed"</string>
-    <string name="sign_in_with_facebook">Sign in with Facebook</string>
-    <string name="sign_in_with_twitter">Sign in with Twitter</string>
-    <string name="sign_in_with_google">Sign in with Google</string>
-    <string name="sign_in">Sign in</string>
-    <string name="forgot_password">Forgot Password?</string>
-    <string name="password">Password</string>
-    <string name="email_of_mobile_number">Email of mobile number</string>
-    <string name="sign_up">Sign up</string>
-    <string name="dont_have_account">Don\'t have an account?</string>
-    <string name="confirm_password">Confirm Password</string>
+
 </resources>
